requirements.txt
async_timeout==4.0.3
asyncpg==0.30.0
drawio==0.0.10
faiss_cpu==1.11.0.post1
fastapi==0.116.1
langchain==0.3.26
numpy==1.21.5
numpy==1.19.5
openai==1.97.0
pgvector==0.4.1
pinecone==7.3.0
psutil==5.9.0
psutil==7.0.0
psycopg2_binary==2.9.10
pydantic==2.11.7
PyPDF2==3.0.1
python-dotenv==1.1.1
rank_bm25==0.2.2
Requests==2.32.4
sentence_transformers==2.3.0
slowapi==0.1.9
torch==2.7.1
transformers==4.35.2
uvicorn==0.35.0
